backend/requirements.txt
Flask
Flask-Cors
google-auth
requests
Flask-SQLAlchemy
flask_migrate
python-dotenv
psycopg2-binary
